Make serial port block, with inter-character timeout
Removing O_NDELAY and O_NONBLOCK lets operations on the file descriptor
(e.g., `read()`) become blocking operations.

Setting `VMIN` to 255 and `VTIME` to 1 puts the file in the *(read with
interbyte timeout)* case.

(see [`termios` manpage](http://man7.org/linux/man-pages/man3/termios.3.html))

This case is well-suited to bursty data (i.e., occasional packets).
Most of our serial sensors provide data in this fashion.

Essentially, with `VMIN = 255` and `VTIME = 1`, a call like:
```c
ssize_t bytes_read = read(fd, buffer, nbyte);
```

will return the first time one of the following conditions is met:

* `nbyte` bytes are available

* `VMIN = 255` bytes are available

* `VTIME = 1` tenths of a second have passed since the last new
  byte became available.

So by changing these parameters in the defaults, we are:

* removing the need for `nanosleep` to prevent sensor daemons
  from spinning up extra CPU cycles to read a few characters at
  a time as they come in

* removing the need for developers to specify a nominal period
  for sensors to output data

at the cost of:

* a *potential* 0.1 second delay

To mitigate that cost:

* `VMIN` could be set to the nominal data length for each sensor, or

* `r2_buffer_fill` could be modified to try the nominal data length.

Another potential solution would be to insert a short `nanosleep` in
`r2_sli_stream`, *after* the `select()`, but *before* the `read()`.
This would have the advantage of being a deterministic delay, which
we could adjust the data timestamp to take into account.

For a 9600 baud sensor with 35-character data frames, we'd want a 3.6
millisecond delay; for a 38400 baud sensor (like the ADCP) with data
frames of (16 + 16 + 952 + 4 = 988) bytes, we'd want a 25 millisecond
delay.